port/wasm/artifact-explorer: remove CSV export feature
There was a glaring bug with this feature -- the mission header was wrong
probably since the beginning (#102 on private tracker). But I never got a bug
report. I guess no one was using the feature. In the off chance that someone
was using the artifact parameters CSV, I'll just point them to the JSON data
instead.