Introduce netlify-headers-expander
Netlify has a nasty edge case where updates are not atomic: index.html is
updated but some of assets referenced are still 404, and thanks to the fact
that caching headers added by _headers are added unconditionally, the 404s are
cached, breaking the site.

My netlify-headers-expander expands glob patterns to existing paths only.