Create colleggtible_eggs.ts
Added functions which will return the colleggtible egg benefit.

To work this will interrogate the backup contracts looking for customEgg's and then using the maxFarmSizeReached query the custom egg backup data and return the appropriate earned benefit based on those egg counts.

Helper functions added to call the main function with the correct parameter for all the GameDimensions. This should be able to handle any collectibles using GameDimensions for buffs.